% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_annotation_list.R
\name{as_annotation_list}
\alias{as_annotation_list}
\title{Convert a data frame to a list of gene sets}
\usage{
as_annotation_list(annotations, set_col, gene_col)
}
\arguments{
\item{annotations}{a data frame containing some annotations}

\item{set_col}{the name of the column that contains the gene sets
(e.g. Gene Ontology terms)}

\item{gene_col}{the name of the column that contains the annotated targets 
(e.g. UniProt accessions)}
}
\value{
a named list where the names are found in keyCol and the entries are 
unique items from termCol
}
\description{
Convert a data frame with gene sets and gene names in separate columns
into a list of gene sets, in which the names of the list are the gene sets
and the entries are vectors of genes.
}
\examples{
# set up a gene set data frame
ann_df = data.frame(pathway = c('a', 'a', 'a', 'b', 'c'),
                    gene = c('gene_1', 'gene_2', 'gene_3', 'gene_1', 
                             'gene_3'))
ann = as_annotation_list(ann_df, 'pathway', 'gene')
str(ann)

# go = read_gpa("goa_human.gpa.gz")
# ann = as_annotation_list(go, "GO ID", "DB Object ID")
}
