% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_complex_auc.R
\name{calculate_complex_auc}
\alias{calculate_complex_auc}
\title{Evaluate the recovery of known protein complexes}
\usage{
calculate_complex_auc(pairs, adj, score_column = "cor")
}
\arguments{
\item{pairs}{a matrix of dimensions (# of proteins) x (# of proteins), 
scoring every possible protein pair, in which higher values reflect more 
similar pairs, e.g. as returned by \link{score_pairs}. Alternatively,
a data frame of candidate protein-protein interactions, with proteins
in the first two columns.}

\item{adj}{an adjacency matrix between all complex proteins, with 
intra-complex pairs as \code{1}s and inter-complex pairs as \code{0}s,
e.g. as returned by \link{to_adjacency_matrix}}

\item{score_column}{when \code{pairs} is a data frame, the column that 
contains the score for each protein pair}
}
\value{
the area under the receiver operating characteristic curve
}
\description{
Assess the intrinsic quality of a CF-MS dataset by evaluating its ability
to recover known protein complexes, using receiver operating characteristic
(ROC) analysis. In this analysis, the correlation coefficients between
every pair of proteins in the dataset are ranked, and compared to a binary
outcome variable reflecting whether the two proteins are in the same complex.
The area under the curve (AUC) is returned as a measure of the ability of the
CF-MS data to recover known complexes. This measure ranges from 0 to 1, 
with 1 representing perfect recovery, and 0.5 representing random recovery.
}
