% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{calculate_features}
\alias{calculate_features}
\title{Calculate and merge features for a list of CF-MS matrices}
\usage{
calculate_features(mats, metric = metrics())
}
\arguments{
\item{mats}{a list of CF-MS matrices, with proteins in rows and fractions in 
columns}

\item{metric}{the measure of association to use in scoring protein pairs}
}
\value{
a data frame containing features for all protein pairs across all 
  replicates
}
\description{
Score each protein pair in a series of CF-MS experiments, using a measure of
association of choice, then merge the calculated features into a single
data frame that contains scores for each pair across all of the input
experiments. This data frame can then be provided to a classifier as input,
alongside a set of 'gold-standard' interacting pairs, in order to score
interactions in a manner that integrates data from multiple CF-MS replicates.
Note that the final data frame may contain missing values; to replace them, 
use the \link{impute_missing_features} function.
}
