% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_matrices.R
\name{combine_matrices}
\alias{combine_matrices}
\title{Combine multiple CF-MS matrices into a single matrix}
\usage{
combine_matrices(mats, mode = c("union", "intersect"))
}
\arguments{
\item{mats}{a list of matrices}

\item{mode}{one of \code{'union'} or \code{'intersect'}; the proteins to
include in the merged matrix}
}
\value{
a merged matrix
}
\description{
Combine fractions from multiple CF-MS matrices into a single matrix, 
containing either the union of all proteins found in any individual matrix,
or the intersection of proteins found in all matrices.
}
