% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matrix.R
\name{filter_matrix}
\alias{filter_matrix}
\title{Filter low-quality chromatograms from a CF-MS chromatogram matrix}
\usage{
filter_matrix(mat, min_fractions = 0)
}
\arguments{
\item{mat}{a CF-MS chromatogram matrix, with proteins in rows and fractions 
in columns}

\item{min_fractions}{minimum number of fractions for a 'good-quality' 
chromatogram}
}
\value{
a filtered matrix with low-quality chromatograms removed
}
\description{
Remove chromatograms from a matrix if the protein in question was not 
quantified in some minimum number of fractions.
}
