% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_pairs.R
\name{filter_pairs}
\alias{filter_pairs}
\title{Filter protein pairs from a chromatogram score matrix}
\usage{
filter_pairs(pairs, mat, min_pairs = 0)
}
\arguments{
\item{pairs}{a square matrix scoring the similarity of each pair of 
chromatograms, as returned by \link{score_pairs}}

\item{mat}{The original CF-MS chromatogram matrix, with proteins in rows and 
fractions in columns}

\item{min_pairs}{minimum number of fractions in which a pair of proteins must 
both be detected to consider scoring the candidate interaction}
}
\value{
a filtered version of the input square matrix, with a subset of 
   protein pairs replaced by \code{NA}s
}
\description{
Replace pairs of proteins in a matrix output by \link{score_pairs} with 
\code{NA}s, if the two proteins were not jointly quantified in some minimum
number of overlapping fractions.
}
