% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_go.R
\name{gaf_colnames}
\alias{gaf_colnames}
\title{Column names for a GAF file}
\usage{
gaf_colnames()
}
\description{
Column names for a tab-delimited annotation file in GAF format. 
Further information about the GAF format is available from the 
\href{http://geneontology.org/page/go-annotation-file-gaf-format-21}{
Gene Ontology Consortium}.
}
