% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_interactions.R
\name{get_splits}
\alias{get_splits}
\title{Split a set of reference protein-protein interactions into cross-validation
folds, for use in \link{score_interactions}.}
\usage{
get_splits(gold_standard, n_folds = 10, split_by = c("proteins", "pairs"))
}
\arguments{
\item{gold_standard}{a data frame with columns \code{protein_A} and 
\code{protein_B}, as returned by \link{to_pairwise_df}, containing 
'gold standard' interacting protein pairs}

\item{n_folds}{the number of folds of cross-validation to perform}

\item{split_by}{the mechanism by which to split the gold standard into
cross-validation folds; either by protein complex subunits 
(\code{'proteins'}) or by pairwise interactions between those subunits 
(\code{'pairs'})}
}
\value{
a list of length \code{n_folds}, in which each item contains 
  \code{"train"} and \code{"test"} data frames with the columns 
  \code{protein_A}, \code{protein_B}, and \code{label}
}
\description{
Split a set of reference protein-protein interactions into cross-validation
folds, for use in \link{score_interactions}.
}
