% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_go.R
\name{go_roots}
\alias{go_roots}
\title{Roots of the Gene Ontology}
\usage{
go_roots()
}
\description{
Return a vector containing the Gene Ontology codes for 'biological process',
'molecular function', and 'cellular compartment'.
}
