% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{impute_missing_features}
\alias{impute_missing_features}
\title{Impute missing features with the median, plus or minus some random noise}
\usage{
impute_missing_features(dat, noise_pct = 0.05)
}
\arguments{
\item{dat}{the feature data frame for which to replace missing data}

\item{noise_pct}{the standard deviation of the random normal 
distribution from which to draw added noise, expressed as a 
percentage of the standard deviation of the non-missing values in each 
column}
}
\value{
a data frame with missing values in each numeric column replaced
by the column median, plus or minus some random noise
}
\description{
Replace missing data within each numeric column of a data frame with 
the column median, plus or minus some random noise, in order to train 
classifiers that do not readily ignore missing data (e.g. random forests or
support vector machines).
}
