% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_pairs.R
\name{label_pairs}
\alias{label_pairs}
\title{Make labels for a list of protein pairs based on a gold standard}
\usage{
label_pairs(gold_standard, dat)
}
\arguments{
\item{gold_standard}{a set of 'gold-standard' interactions. Can be provided
in the form of an adjacency matrix, a data frame of protein pairs, or
a list of protein complexes}

\item{dat}{a data frame with interacting proteins in the first two
columns}
}
\value{
a vector of the same length as the number of rows in \code{dat}, 
  containing \code{NA}s for protein pairs not in the gold standard, and
  ones or zeroes otherwise
}
\description{
Create labels for a data frame containing protein pairs, by cross-referencing
those protein pairs with a 'gold standard' set of known interactions. 
The gold standard interactions can be provided in a variety of different
formats.
}
