% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_features.R
\name{merge_features}
\alias{merge_features}
\title{Merge features across multiple replicates}
\usage{
merge_features(feature_dfs)
}
\arguments{
\item{feature_dfs}{a list of feature data frames, each of which has protein
pairs in the first two columns. Alternatively, a list of square matrices 
can be provided as input, and will be coerced into a list of feature
data frames.}
}
\value{
a data frame containing features for all protein pairs across all 
  replicates
}
\description{
Merge features extracted from multiple replicates into a single data 
frame for input to a classifier. Note that the merged data frame may contain 
missing values; to replace them, use the \link{impute_missing_features}
function.
}
