% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Metrics available in the CFTK}
\usage{
metrics()
}
\value{
a vector enumerating the measures of association implemented in CFTK
}
\description{
Returns a named vector of measures of association available for scoring 
chromatogram similarity in CFTK, where the names of the vector represent
human-readable names and the entries themselves are inputs to 
\link{score_pairs}.
}
