% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pr.R
\name{plot_pr}
\alias{plot_pr}
\title{Plot a precision-recall curve for a list of protein-protein interactions}
\usage{
plot_pr(pairs, max_n = NULL)
}
\arguments{
\item{pairs}{a data frame in which each row represents a protein pair, and
  which contains the following columns:
\enumerate{
  \item \code{score}: the score assigned to that protein pair, e.g., by a 
    machine-learning classifier, in which higher scores represent a greater 
    probability of a physical interaction
  \item \code{label}: a ground-truth annotation of whether that pair of
    proteins is known to physically interact; one of \code{1}, \code{0}, or 
    \code{NA} (not labelled) 
}}

\item{max_n}{optionally, plot the precision-recall curve only out to this
maximum number of interactions}
}
\value{
a \code{ggplot2} object plotting the precision-recall curve for the input 
  interactions
}
\description{
Assess the intrinsic quality of a CF-MS dataset by evaluating the 
precision of the interaction network for any number of interactions scored. 
The x-axis in the output plot represents the total number of interactions 
(and can optionally be subset to a limited range), while the y-axis shows
the precision for each point in the ranked list.
}
\details{
Note that the x-axis
shows recall in terms of the total number of interactions in the CF-MS 
dataset, not the proportion of true-positive interactions recovered from the
'gold standard' dataset, since many of these interactions may not have been
detected by CF-MS.
}
