% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{Plot a ROC curve for a list of protein-protein interactions}
\usage{
plot_roc(pairs)
}
\arguments{
\item{pairs}{a data frame in which each row represents a protein pair, and
  which contains the following columns:
\enumerate{
  \item \code{score}: the score assigned to that protein pair, e.g., by a 
    machine-learning classifier, in which higher scores represent a greater 
    probability of a physical interaction
  \item \code{label}: a ground-truth annotation of whether that pair of
    proteins is known to physically interact; one of \code{1}, \code{0}, or 
    \code{NA} (not labelled) 
}}
}
\value{
a \code{ggplot2} object plotting the ROC curve for the input 
  interactions
}
\description{
Assess the intrinsic quality of a CF-MS dataset by evaluating its ability
to recover known protein complexes, using receiver operating characteristic
(ROC) analysis. This function visualizes the trade-off between the proportion 
of true-positive interactions recovered at any given rate of false-positive 
interactions. In the ideal analysis, this trade-off is skewed towards the top
left of the plot area, recovering all true positives without any false 
positives. The dashed line shows the performance of a theoretical random 
classifier.
}
