% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_corum.R
\name{read_corum}
\alias{read_corum}
\title{Read complexes from the CORUM database}
\usage{
read_corum(filepath, identifiers = c("uniprot", "entrez", "gene_name"))
}
\arguments{
\item{filepath}{path to the 'coreComplexes.txt' file}

\item{identifiers}{the set of protein identifiers to parse from the file;
one of \code{'uniprot'}, \code{'entrez'}, or \code{'gene_name'}}
}
\value{
a data frame with protein complex names in column 1 and individual
  subunits in column 2
}
\description{
Utility function to read complexes from the 'coreComplexes.txt' file 
distributed through the CORUM database into a two-column data frame, mapping
each complex (column 1) to its subunits (column 2).
}
