% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_go.R
\name{read_gaf}
\alias{read_gaf}
\title{Read a GAF file}
\usage{
read_gaf(
  filepath,
  filter_NOT = TRUE,
  filter_evidence = NULL,
  ontology = NULL,
  propagate = TRUE
)
}
\arguments{
\item{filepath}{the location of the GAF file}

\item{filter_NOT}{if true, filter annotations with the qualifier NOT}

\item{filter_evidence}{optionally, specify evidence codes to filter. By 
default, evidence codes ND, IPI, IEA and NAS are filtered.}

\item{ontology}{optionally, provide the Gene Ontology itself as an object
read in by the \code{ontologyIndex} package from an OBO file}

\item{propagate}{if true, and an ontology file is provided, all ancestors of 
a given term are associated with each protein.}
}
\value{
a data.frame containing the filtered GAF file
}
\description{
Read a GO annotation file in GAF format. Further information about the GAF
format is available from the 
\href{http://geneontology.org/page/go-annotation-file-gaf-format-21}{Gene Ontology Consortium}.
}
\examples{
# read human GOA and filtering IEA annotations
# go = read_gaf("goa_human.gaf.gz", filter_evidence = "IEA")
# read mouse GOA wthout filtering any annotations
# go = read_gaf("goa_mouse.gaf.gz", filter_evidence = NULL)
}
