% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_maxquant.R
\name{read_maxquant}
\alias{read_maxquant}
\title{Read CF-MS data from MaxQuant}
\usage{
read_maxquant(
  filepath,
  identifiers = c("protein groups", "genes"),
  quant_mode = c("iBAQ", "MS1 intensity", "MaxLFQ", "spectral counts", "ratio")
)
}
\arguments{
\item{filepath}{path to the proteinGroups.txt file}

\item{identifiers}{one of \code{'protein groups'} or \code{'genes'}; if the
latter, protein groups will be matched to gene names (using the 
'Gene names' column in the proteinGroups file), keeping only the best row
per gene when there are redundant gene names)}

\item{quant_mode}{the protein quantitations to use; one of \code{'iBAQ'}, 
\code{'MS1 intensity'}, \code{'MaxLFQ'}, \code{'spectral counts'}, or 
\code{'ratio'}. Note that in ratiometric experiments with more than one
isotopic label, the output matrix may need to be manually split further
(e.g. to separate heavy and medium over light ratios).}
}
\value{
a numeric matrix with proteins in rows and fractions in columns,
  and a \code{'metadata'} attribute containing metadata from the 
  proteinGroups file for each row
}
\description{
Read in a 'proteinGroups.txt' file output by MaxQuant, and convert it to a 
CF-MS chromatogram matrix, in which each protein is a row and each fraction 
is a column.
}
