% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_values.R
\name{replace_missing_values}
\alias{replace_missing_values}
\title{Replace missing values in a CF-MS chromatogram matrix}
\usage{
replace_missing_values(
  mat,
  missing = c("NA", "zero", "noise"),
  noise_floor = 0.001
)
}
\arguments{
\item{mat}{a CF-MS chromatogram matrix, with proteins in rows and fractions 
in columns}

\item{missing}{one of \code{'NA'}, \code{'zero'}, or \code{'noise'}. 
If \code{'zero'}, missing values will be replaced with zeroes.
If \code{'noise'}, missing values will be replaced with random noise, 
sampled from a uniform distribution with a maximum value of 
\code{noise_floor}. If \code{'NA'}, missing values will be left as 
\code{NA}s.}

\item{noise_floor}{upper limit of the uniform distribution for 
\code{missing = 'noise'}}
}
\value{
a copy of the input matrix with missing values optionally replaced
}
\description{
Sanitize different representations of missing values (e.g., zeroes, 
\code{NaN}s, or infinite values) in a chromatogram matrix, then optionally
replace them all with zeroes or with near-zero noise.
}
\examples{
mat = matrix(rnorm(100), ncol = 10, nrow = 10,
             dimnames = list(paste0('protein_', LETTERS[1:10]),
                             paste0('fraction_', seq_len(10))))
mat[sample(seq_along(mat), 5)] = 0
mat[sample(seq_along(mat), 8)] = NA
mat[sample(seq_along(mat), 2)] = Inf
sanitized = replace_missing_values(mat, missing = 'zero')
table(sanitized == 0)

}
