% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_values.R
\name{sanitize_missing_values}
\alias{sanitize_missing_values}
\title{Sanitize missing values in a CF-MS chromatogram matrix}
\usage{
sanitize_missing_values(mat)
}
\arguments{
\item{mat}{a CF-MS chromatogram matrix, with proteins in rows and fractions 
in columns}
}
\value{
a sanitized matrix with all zeroes, infinite values, and \code{NaN}s
  replaced with \code{NA}
}
\description{
Standardize the way missing values are represented in a chromatogram matrix
by replacing zeroes, infinite values, and \code{NaN}s with \code{NA}.
}
