% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_interactions.R
\name{score_interactions}
\alias{score_interactions}
\title{Score potential interactions in cross-validation}
\usage{
score_interactions(
  features,
  gold_standard,
  classifier = c("RF", "NB", "SVM", "LR"),
  split_by = c("proteins", "pairs"),
  n_folds = 10,
  verbose = TRUE
)
}
\arguments{
\item{features}{a data frame containing features for all protein pairs 
across all replicates, containing columns \code{protein_A} and 
\code{protein_B}, as returned by \link{calculate_features}}

\item{gold_standard}{a data frame with columns \code{protein_A} and 
\code{protein_B}, as returned by \link{to_pairwise_df}, containing 
'gold standard' interacting protein pairs}

\item{classifier}{the classifier to use; one of \code{'RF'} (random forest),
\code{'NB'} (naive Bayes), \code{'SVM'} (support vector machine), or
\code{'LR'} (logistic regression)}

\item{split_by}{the mechanism by which to split the gold standard into
cross-validation folds; either by protein complex subunits 
(\code{'proteins'}) or by pairwise interactions between those subunits 
(\code{'pairs'})}

\item{n_folds}{the number of folds of cross-validation to perform}

\item{verbose}{set to \code{FALSE} to disable messages from the function}
}
\description{
Use a machine-learning approach to integrate data from across multiple CF-MS
replicates, or multiple features within a single replicate. This function
takes a data frame of features as input, alongside a set of 'gold-standard'
reference interactions. The gold standard is split into a user-specified
number of folds, and a classifier is trained on the reference interactions
after leaving out each fold in turn. Each classifier is then applied to
predict interactions in the entire feature data frame, minus the protein
pairs that overlap with the training interactions for that fold. The mean
classifier score across all folds is calculated for each protein pair,
and the proteins are sorted in descending order by their mean score.
}
