% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pairs.R
\name{score_pairs}
\alias{score_pairs}
\title{Score the similarity of all pairs of chromatograms}
\usage{
score_pairs(mat, metric = metrics(), seed = 0)
}
\arguments{
\item{mat}{a CF-MS chromatogram matrix, with proteins in rows and fractions 
in columns}

\item{metric}{the measure of association to use in scoring protein pairs}

\item{seed}{seed for the random number generator, used to ensure the 
reproducibility some metrics, such as GENIE3}
}
\value{
a matrix of dimensions (# of proteins) x (# of proteins), scoring
  every possible protein pair, in which higher values reflect more similar
  pairs
}
\description{
Compute an index measuring the similarity of chromatogram profiles for every 
pair of proteins detected in a CF-MS experiment. This function implements
a total of 24 such indices, or measures of association. The full list 
can be accessed using the \link{metrics} function.
}
