% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_precision.R
\name{threshold_precision}
\alias{threshold_precision}
\title{Threshold interactions at a given precision cutoff}
\usage{
threshold_precision(interactions, threshold)
}
\arguments{
\item{interactions}{the ranked list of interactions output by
\code{\link{score_interactions}}, including a \code{precision}
column as calculated after the fact by \code{\link{calculate_precision}}}

\item{threshold}{the minimum precision of the unweighted interaction
network to return}
}
\value{
the subset of the original ranked list of interactions including 
  the greatest number of interactions that yields a network at the given
  precision
}
\description{
Threshold interactions at a given precision cutoff
}
