% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pairwise_df.R
\name{to_pairwise_df}
\alias{to_pairwise_df}
\title{Convert an annotations list to a pairwise data frame}
\usage{
to_pairwise_df(ann)
}
\arguments{
\item{ann}{a list of gene sets, e.g. as returned by \link{as_annotation_list}}
}
\value{
a data frame with two columns, \code{'protein_A'} and 
  \code{'protein_B'}, containing all unique pairs of protiens found within 
  the same gene set
}
\description{
Convert a list of gene sets (for instance, where each entry corresponds to a
complex and contains all the protein subunits of that complex) into a data
frame enumerating all of the possible pairs within each list item.
}
