% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_matrix.R
\name{transform_matrix}
\alias{transform_matrix}
\title{Log-transform or quantile normalize a CF-MS matrix}
\usage{
transform_matrix(mat, mode = c("quantile", "log-transform"))
}
\arguments{
\item{mat}{a CF-MS chromatogram matrix, with proteins in rows and fractions 
in columns}

\item{mode}{the transformation to apply; one of \code{'quantile'} or 
\code{'log-transform'}}
}
\description{
Transform a CF-MS matrix by either log-transformation or quantile 
normalization. Note that quantile normalization relies on the implementation
in the \code{preprocessCore} Bioconductor package, which is not installed 
alongside the CFTK pacakge, and must be installed from Bioconductor 
separately by the user.
}
